kmm gradle config tidy up
