+++ buildSrc/src/main/kotlin/buildsrc/convention/kotlin-multiplatform.gradle.kts
@@ -2,6 +2,7 @@ package buildsrc.convention
 
 import buildsrc.config.relocateKotlinJsStore
 
+
 plugins {
   id("buildsrc.convention.subproject")
   kotlin("multiplatform")
@@ -10,3 +11,15 @@ plugins {
 
 
 relocateKotlinJsStore()
+
+
+kotlin {
+  targets.all {
+    compilations.all {
+      kotlinOptions {
+        languageVersion = "1.6"
+        apiVersion = "1.6"
+      }
+    }
+  }
+}
