+++ modules/kxs-ts-gen-core/build.gradle.kts
@@ -1,3 +1,6 @@
+import buildsrc.config.publicationsFromMainHost
+
+
 plugins {
   buildsrc.convention.`kotlin-multiplatform`
   buildsrc.convention.`maven-publish`
@@ -8,14 +11,6 @@ plugins {
 val kotlinxSerializationVersion = "1.3.2"
 
 kotlin {
-//  val hostOs = System.getProperty("os.name")
-//  val isMingwX64 = hostOs.startsWith("Windows")
-//  val nativeTarget = when {
-//    hostOs == "Mac OS X" -> macosX64("native")
-//    hostOs == "Linux"    -> linuxX64("native")
-//    isMingwX64           -> mingwX64("native")
-//    else                 -> throw GradleException("Host OS is not supported in Kotlin/Native.")
-//  }
 
 //  js(IR) {
 //    binaries.executable()
@@ -25,12 +20,11 @@ kotlin {
 //      }
 //    }
 //  }
+
   jvm {
     compilations.all {
       kotlinOptions {
         jvmTarget = "11"
-        languageVersion = "1.6"
-        apiVersion = "1.6"
       }
     }
     withJava()
@@ -38,6 +32,18 @@ kotlin {
       useJUnitPlatform()
     }
   }
+
+//  publishing {
+//    publications {
+//      matching { it.name in publicationsFromMainHost() }.all {
+//        val targetPublication = this@all
+//        tasks.withType<AbstractPublishToMaven>()
+//          .matching { it.publication == targetPublication }
+//          .configureEach { onlyIf { findProperty("isMainHost") == "true" } }
+//      }
+//    }
+//  }
+
   sourceSets {
 
     all {
