+++ settings.gradle.kts
@@ -9,3 +9,8 @@ include(
 )
 
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+
+dependencyResolutionManagement {
+  @Suppress("UnstableApiUsage") // Central declaration of repositories is an incubating feature
+  repositoriesMode.set(RepositoriesMode.PREFER_SETTINGS)
+}
